requirements.txt
streamlit
pandas
plotly
numpy
searoute
geopy
Pillow